adds async await

+++ controllers/feed.js
@@ -4,67 +4,72 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
-exports.getPosts = (req, res, next) => {
+exports.getPosts = async (req, res, next) => {
   const currentPage = req.query.page || 1;
   const perPage = 2;
   let totalItems;
-  Post.find()
-    .countDocuments()
-    .then(count => {
-      totalItems = count;
-      return Post.find()
-        .skip((currentPage - 1) * perPage)
-        .limit(perPage)
-        .then(posts => {
-          res
-            .status(200)
-            .json({ message: 'Fetched posts succesfully', posts: posts, totalItems });
-        })
-    })
-    .catch(err => {
-      if (error.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
+  try {
+    const totalItems = await Post.find().countDocuments();
+    const posts = await Post.find()
+      .skip((currentPage - 1) * perPage)
+      .limit(perPage);
+
+    res.status(200).json({
+      message: 'Fetched posts successfully.',
+      posts: posts,
+      totalItems: totalItems
     });
+  } catch (error) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+    next(err);
+  }
 };
 
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error('Validation failed, data incorrect');
+    const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
     throw error;
   }
   if (!req.file) {
-    const error = new Error('No image required');
+    const error = new Error('No image provided.');
     error.statusCode = 422;
     throw error;
   }
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
-    creator: {
-      name: 'Nico'
-    },
     imageUrl: imageUrl,
-    creator: { name: 'Nico' }
+    creator: req.userId
   });
   post
     .save()
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
-        message: 'post created successfully!',
-        post: result
+        message: 'Post created successfully!',
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
-      if (error.statusCode) {
+      if (!err.statusCode) {
         err.statusCode = 500;
       }
       next(err);
@@ -72,11 +77,11 @@ exports.createPost = (req, res, next) => {
 };
 
 exports.getPost = (req, res, next) => {
-  postId = req.params.postId;
+  const postId = req.params.postId;
   Post.findById(postId)
     .then(post => {
       if (!post) {
-        const error = new Error('Could not find post!');
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
@@ -84,17 +89,17 @@ exports.getPost = (req, res, next) => {
     })
     .catch(err => {
       if (!err.statusCode) {
-        err.statuscode = 500;
+        err.statusCode = 500;
       }
       next(err);
-    })
+    });
 };
 
 exports.updatePost = (req, res, next) => {
   const postId = req.params.postId;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error('Validation failed, data incorrect');
+    const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
     throw error;
   }
@@ -105,15 +110,20 @@ exports.updatePost = (req, res, next) => {
     imageUrl = req.file.path;
   }
   if (!imageUrl) {
-    const error = new Error('No file picked!');
+    const error = new Error('No file picked.');
     error.statusCode = 422;
     throw error;
   }
   Post.findById(postId)
     .then(post => {
       if (!post) {
-        const error = new Error('Could not find post!');
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -132,26 +142,36 @@ exports.updatePost = (req, res, next) => {
         err.statusCode = 500;
       }
       next(err);
-    })
+    });
 };
-
 
 exports.deletePost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
     .then(post => {
       if (!post) {
-        const error = new Error('Could not find post!');
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
-      // check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
       clearImage(post.imageUrl);
-      return Post.findByIdAndRemove(postId)
+      return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
-      res.status(200).json({ message: 'Deleted post!' });
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: 'Deleted post.' });
     })
     .catch(err => {
       if (!err.statusCode) {
